+++ lib/ui/save_states/save_states_screen.dart
@@ -2,12 +2,14 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:nesd/ui/common/confirmation_dialog.dart';
 import 'package:nesd/ui/common/focus_child.dart';
 import 'package:nesd/ui/common/nesd_menu_wrapper.dart';
-import 'package:nesd/ui/emulator/main_menu/recent_rom_list.dart';
+import 'package:nesd/ui/common/rom_list.dart';
 import 'package:nesd/ui/emulator/nes_controller.dart';
 import 'package:nesd/ui/emulator/rom_manager.dart';
 import 'package:nesd/ui/router.dart';
+import 'package:nesd/ui/save_states/save_states_screen_controller.dart';
 import 'package:path/path.dart' as p;
 
 @RoutePage()
@@ -18,15 +20,10 @@ class SaveStatesScreen extends HookConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final controller = ref.watch(nesControllerProvider);
-    final romManager = ref.watch(romManagerProvider);
+    final nesController = ref.watch(nesControllerProvider);
+    final controller = ref.watch(saveStatesScreenControllerProvider(romInfo));
 
-    final romsFuture = useMemoized(
-      () => _getRomTileDataForRom(romManager, romInfo),
-      [romInfo],
-    );
-
-    final romsSnapshot = useFuture(romsFuture);
+    final romsSnapshot = useStream(controller.stream);
 
     if (romsSnapshot.hasError) {
       return const Center(child: Text('Error loading ROMs'));
@@ -37,6 +34,21 @@ class SaveStatesScreen extends HookConsumerWidget {
     }
 
     final roms = romsSnapshot.data!;
+
+    Future<void> delete(BuildContext context, RomTileData romTileData) async {
+      final confirmed = await ConfirmationDialog.show(
+        context,
+        title: const Text('Delete save state?'),
+        content: Text(
+          'Are you sure you want to delete'
+          ' the save state ${romTileData.title}?',
+        ),
+      );
+
+      if (confirmed == true) {
+        controller.delete(romTileData);
+      }
+    }
 
     return Scaffold(
       appBar: AppBar(
@@ -55,35 +67,29 @@ class SaveStatesScreen extends HookConsumerWidget {
             child: RomList(
               roms: roms,
               onPressed: (romTileData) {
-                controller.loadRom(
+                nesController.loadRom(
                   romTileData.romInfo.path,
                   state: romTileData.state,
                 );
                 ref.read(routerProvider).navigate(const MainRoute());
               },
+              onRemove:
+                  (romTileData) async => await delete(context, romTileData),
+              contextMenuBuilder:
+                  (context, romTileData, close) => [
+                    ListTile(
+                      title: const Text('Delete save state'),
+                      onTap: () async {
+                        close();
+
+                        await delete(context, romTileData);
+                      },
+                    ),
+                  ],
             ),
           ),
         ),
       ),
     );
-  }
-
-  Future<List<RomTileData>> _getRomTileDataForRom(
-    RomManager romManager,
-    RomInfo romInfo,
-  ) async {
-    final romTileDatas = <RomTileData>[];
-
-    for (var slot = 0; slot < 10; slot++) {
-      final romTileData = await romManager.getRomTileDataForSlot(romInfo, slot);
-
-      if (romTileData == null) {
-        continue;
-      }
-
-      romTileDatas.add(romTileData);
-    }
-
-    return romTileDatas;
   }
 }